[feature|build]在大屏幕下，首页右侧会显示表情包信息；更新依赖版本；更新Gradle版本

+++ app/src/main/java/com/skyd/rays/ui/screen/home/HomeState.kt
@@ -24,5 +24,5 @@ sealed class SearchResultUiState {
 
 sealed class PopularTagsUiState {
     object Init : PopularTagsUiState()
-    data class Success(val popularTags: List<Pair<String, Long>>) : PopularTagsUiState()
+    data class Success(val popularTags: List<Pair<String, Float>>) : PopularTagsUiState()
 }
